Desktop/A_Tuo_Servizio/services/servizio_portfolio/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
pytest
httpx # Required by TestClient
